Improve SDK initialisation
Restructures the wrapper's interface to allow for two modes of usage;
as an internally-managed singleton through the exported functions
`Configure`, `Connect`, and `GetDefaultClient`, or as a self-managed
object by calling the `NewClient` constructor.

Also separates the configuration and connection phases by introducing
a `Connect` method. This enables consumers greater control over
when configuration is supplied to the SDK and when a live connection
is established.