doc: add MIT license
This repo is public, and requires a license to be properly recognised by the Go tooling (for purposes like documentation generation). The MIT license is sufficient here: there's no need for the code in this repo to be encumbered with the reciprocal clauses provided by the GPL or similar licenses.