fix: set a default init wait time (#11)
Sets a default initWait value of 5 seconds so the client is able to
block (for up to 5 seconds) for the initial connection to LD to succeed.

This was not well documented and led to issues when users attempted to
query flags with an uninitialised client. Setting a default value here
simplifies the happy path. The option always exists to override the
wait time with a custom value.