feat: allow the LD client to be configured in test mode (#23)
* feat: allow client to be initialised in test mode

Adds a WithTestMode() ConfigOption to configure the wrapper in test
mode. Test data can be supplied at runtime using the existing functions
on the LD SDK, or via a local JSON file.

* refactor: organise client tests

* feat: source test data from .ld-flags.json if exists

This means that if LAUNCHDARKLY_CONFIGURATION is not set, but an
.ld-flags.json file exists in the directory the Go program was invoked
from, the SDK will automatically fall-back to test mode using the JSON
file as the data source.

This provides a transparent way for users to consume the SDK for
local development or testing. They don't need to modify the
initialisation code to explicitly invoke the test mode.